Create style.css, and add how to examples on files
Completed the less from the Odin project and took
notes within the style.css and index.html using
placeholder universal, type, class, and ID selectors
with different case scenarios for each to ensure I
have all the necessary information to complete the
project.

Continuing on to complete the assignment portion
by creating properties to match with the
desired-outcome.png.