Add flex properties to center container to page
Upon furthe research and understanding, in order
to make the flexbox centered to the page dynamically
a third div/container/class needed to be made in
the index.html file. Upon interacting with the
properties of this class in the style.css file,
I was able to center this box on the page using
the same commands on the container for the red box.

Now moving on to the next exercise after learning
everything I needed to know for this one.