Add correct properties to match desired png
After going through each possible use of diplay
flex, I was able to get far enough to have created
flexboxes for each of the selectors that need to be
grouped or separated. Focused too much on using only
flex properties which did confuse me at the end but,
after checking over the solution css file I was able
to understand how the properties can be configured
efficiently.

Moving on to next exercise.