Copy solution answers to explain exercise better
Was unable to solve this exercise, unsure if I need
to go back over readings or need more hands on practice.
Very confused on when to use display flex or not to,
the containerization is also getting a little confusing.

Continuing on to next exerise to see if things get
any more familiar or click for me later.