Add all necessary properties to webpage files
In order to complete this final exercise for flexbox,
a deep understanding of how flex containers work and
are able to interact with other flex containers.
By ensuring every section of the webpage is apart
of a container, I was able to correctly section off
the header, the content, and the footer to match
with the desired png. Thus, I have created a dynamic
flex layout using all the correct properties.

Moving on to the next section of TOP.